Moving state machine related types to their own files
