+++ src/Kinetic/Linq/StateMachines/IStateMachine.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace Kinetic.Linq.StateMachines;
+
+public interface IStateMachine<T> : IObserver<T>, IDisposable
+{
+    StateMachineBox Box { get; }
+
+    void Initialize(StateMachineBox box);
+}
